created a reusable login form
